[Chrome Next] Revert navigation package to main
Reset @kbn/core-chrome-navigation to upstream main, reverting all
sidenav-specific additions: ToolItem renderContent/renderPopover,
IconButton component, HeaderToolbar/FooterToolbar, tool shortcuts,
and related types/stories/tests.
