+++ src/core/packages/chrome/navigation/index.ts
@@ -17,6 +17,4 @@ export type {
   SecondaryMenuItem,
   SecondaryMenuSection,
   SideNavLogo,
-  ToolItem,
-  ToolSlots,
 } from './types';
