+++ src/core/packages/chrome/navigation/src/__stories__/navigation.stories.tsx
@@ -9,16 +9,7 @@
 
 import React, { useState } from 'react';
 import type { ComponentProps } from 'react';
-import {
-  EuiAvatar,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiHorizontalRule,
-  EuiLink,
-  EuiSkipLink,
-  EuiText,
-  useEuiTheme,
-} from '@elastic/eui';
+import { EuiSkipLink, useEuiTheme } from '@elastic/eui';
 import type { UseEuiTheme } from '@elastic/eui';
 import type { Meta, StoryFn, StoryObj } from '@storybook/react';
 import { APP_MAIN_SCROLL_CONTAINER_ID } from '@kbn/core-chrome-layout-constants';
@@ -50,188 +41,6 @@ const styles = ({ euiTheme }: UseEuiTheme) => {
 };
 
 type PropsAndArgs = ComponentProps<typeof Navigation>;
-
-const CHROME_TOOLS: NonNullable<PropsAndArgs['tools']> = {
-  headerTools: [
-    {
-      id: 'search',
-      label: 'Search',
-      iconType: 'search',
-      onClick: () => {},
-    },
-    {
-      id: 'notifications',
-      label: 'Notifications',
-      iconType: 'bell',
-      onClick: () => {},
-    },
-    {
-      id: 'quick_settings',
-      label: 'Quick settings',
-      iconType: 'gear',
-      onClick: () => {},
-    },
-  ],
-  footerTools: [
-    {
-      id: 'help',
-      label: 'Help',
-      iconType: 'question',
-      sections: [
-        {
-          id: 'help-links',
-          items: [{ id: 'documentation', label: 'Documentation', href: '/help/documentation' }],
-        },
-      ],
-    },
-  ],
-};
-
-const SpaceBadge = () => (
-  <span
-    css={css`
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      width: 20px;
-      height: 20px;
-      border-radius: 4px;
-      background-color: #f04e98;
-      color: #fff;
-      font-size: 11px;
-      font-weight: 700;
-      line-height: 1;
-    `}
-  >
-    D
-  </span>
-);
-
-const SPACES = [
-  { id: 'default', name: 'Default', color: '#f04e98', initial: 'D' },
-  { id: 'marketing', name: 'Marketing', color: '#00bfb3', initial: 'M' },
-  { id: 'engineering', name: 'Engineering', color: '#6092c0', initial: 'E' },
-  { id: 'sales', name: 'Sales', color: '#d6bf57', initial: 'S' },
-];
-
-const SpacePicker = ({ closePopover }: { closePopover: () => void }) => (
-  <EuiFlexGroup
-    direction="column"
-    gutterSize="none"
-    css={css`
-      padding: 12px;
-    `}
-  >
-    <EuiFlexItem>
-      <EuiText
-        size="xs"
-        css={css`
-          font-weight: 600;
-          margin-bottom: 8px;
-        `}
-      >
-        <p>Spaces</p>
-      </EuiText>
-    </EuiFlexItem>
-    {SPACES.map((space) => (
-      <EuiFlexItem key={space.id}>
-        <button
-          type="button"
-          onClick={closePopover}
-          css={css`
-            display: flex;
-            align-items: center;
-            gap: 8px;
-            padding: 6px 4px;
-            border: none;
-            background: transparent;
-            cursor: pointer;
-            width: 100%;
-            border-radius: 4px;
-            &:hover {
-              background: rgba(0, 0, 0, 0.05);
-            }
-          `}
-        >
-          <span
-            css={css`
-              display: inline-flex;
-              align-items: center;
-              justify-content: center;
-              width: 18px;
-              height: 18px;
-              border-radius: 3px;
-              background-color: ${space.color};
-              color: #fff;
-              font-size: 10px;
-              font-weight: 700;
-              flex-shrink: 0;
-            `}
-          >
-            {space.initial}
-          </span>
-          <EuiText size="s">{space.name}</EuiText>
-        </button>
-      </EuiFlexItem>
-    ))}
-    <EuiFlexItem>
-      <EuiHorizontalRule margin="s" />
-      <EuiLink
-        href="/spaces"
-        onClick={(e) => {
-          e.preventDefault();
-          closePopover();
-        }}
-      >
-        <EuiText size="xs">Manage spaces</EuiText>
-      </EuiLink>
-    </EuiFlexItem>
-  </EuiFlexGroup>
-);
-
-const CUSTOM_TOOLS: NonNullable<PropsAndArgs['tools']> = {
-  headerTools: [
-    {
-      id: 'spaces',
-      label: 'Current space',
-      renderContent: () => <SpaceBadge />,
-      renderPopover: (closePopover) => <SpacePicker closePopover={closePopover} />,
-      popoverWidth: 360,
-    },
-    { id: 'search', label: 'Search', iconType: 'search', onClick: () => {} },
-  ],
-  footerTools: [
-    {
-      id: 'help',
-      label: 'Help',
-      iconType: 'question',
-      sections: [
-        {
-          id: 'help-links',
-          items: [{ id: 'docs', label: 'Documentation', href: '/docs' }],
-        },
-      ],
-    },
-    {
-      id: 'user',
-      label: 'Jane Doe',
-      renderContent: () => <EuiAvatar size="s" name="Jane Doe" />,
-      sections: [
-        {
-          id: 'account',
-          items: [
-            { id: 'profile', label: 'Profile', href: '/profile' },
-            { id: 'preferences', label: 'Preferences', href: '/preferences' },
-          ],
-        },
-        {
-          id: 'session',
-          items: [{ id: 'logout', label: 'Log out', href: '/logout' }],
-        },
-      ],
-    },
-  ],
-};
 
 const PreventLinkNavigation = (Story: StoryFn) => {
   usePreventLinkNavigation();
@@ -275,31 +84,6 @@ export const Default: StoryObj<PropsAndArgs> = {
       );
     },
   ],
-  render: (args) => <ControlledNavigation {...args} />,
-};
-
-export const WithTools: StoryObj<PropsAndArgs> = {
-  name: 'With header and footer tools',
-  decorators: [
-    (Story) => {
-      return (
-        <>
-          <Global styles={styles} />
-          <Story />
-        </>
-      );
-    },
-  ],
-  args: {
-    tools: CHROME_TOOLS,
-    logo: {
-      id: 'observability',
-      href: LOGO.href,
-      label: LOGO.label,
-      iconType: LOGO.iconType,
-      hideLabel: true,
-    },
-  },
   render: (args) => <ControlledNavigation {...args} />,
 };
 
@@ -360,31 +144,6 @@ export const WithManyItems: StoryObj<PropsAndArgs> = {
         },
       ],
       footerItems: PRIMARY_MENU_FOOTER_ITEMS,
-    },
-  },
-  render: (args) => <ControlledNavigation {...args} />,
-};
-
-export const WithCustomToolItems: StoryObj<PropsAndArgs> = {
-  name: 'With custom tool items (avatar & space badge)',
-  decorators: [
-    (Story) => {
-      return (
-        <>
-          <Global styles={styles} />
-          <Story />
-        </>
-      );
-    },
-  ],
-  args: {
-    tools: CUSTOM_TOOLS,
-    logo: {
-      id: 'observability',
-      href: LOGO.href,
-      label: LOGO.label,
-      iconType: LOGO.iconType,
-      hideLabel: true,
     },
   },
   render: (args) => <ControlledNavigation {...args} />,
