+++ src/core/packages/chrome/navigation/src/__tests__/both_modes.test.tsx
@@ -1251,7 +1251,6 @@ describe('Both modes', () => {
         const serviceInventoryLink = within(sidePanel).getByTestId(
           sidePanelItemId('service-inventory')
         );
-        const collapseButton = screen.getByTestId('sideNavCollapseButton');
 
         act(() => {
           solutionLogo.focus();
@@ -1269,10 +1268,7 @@ describe('Both modes', () => {
 
         expect(gettingStartedLink).toHaveFocus();
 
-        await user.tab();
-
-        expect(collapseButton).toHaveFocus();
-
+        // Tab to side panel - should land on first focusable item (Service inventory)
         await user.tab();
 
         expect(serviceInventoryLink).toHaveFocus();
