+++ src/core/packages/chrome/navigation/src/utils/get_has_submenu.ts
@@ -7,15 +7,14 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+import type { MenuItem } from '../../types';
+
 /**
  * Utility function for checking whether the menu item has a submenu.
  *
  * @param item - the menu item to check.
  * @returns `true` if the menu item has a submenu, `false` otherwise.
  */
-export const getHasSubmenu = (item: {
-  sections?: ReadonlyArray<unknown>;
-  renderPopover?: unknown;
-}): boolean => {
-  return (!!item.sections && item.sections.length > 0) || !!item.renderPopover;
+export const getHasSubmenu = (item: MenuItem): boolean => {
+  return !!item.sections && item.sections.length > 0;
 };
